Create a CashSystem.
This commit moves the Cash System from the GameManager into its own class so you can use it in multiple scripts without having to rewrite the same methods over again.
