+++ Assets/Jacob/Scripts/Data/CashSystem.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3f9572d0625d4f00a60fec5a24501bd6
+timeCreated: 1680203790
